First implementation of a PSO for a test objective function

+++ src/estimation/PSO_Best_Particle.h
@@ -13,19 +13,19 @@ class PSO_Best_Particle{
 public:
     // Constructor
     PSO_Best_Particle();
-    PSO_Best_Particle(std::vector<int> position, int fitness);
+    PSO_Best_Particle(std::vector<float> position, float fitness);
     //PSO_Best_Particle(const PSO_Particle &particle);
 
     // Setters
-    void set_position(std::vector<int> position);
-    void set_fitness(int fitness);
+    void set_position(std::vector<float> position);
+    void set_fitness(float fitness);
 
     // Getters
-    std::vector<int> get_position();
-    int get_fitness();
+    std::vector<float> get_position();
+    float get_fitness();
 private:
-    std::vector<int> position;
-    int fitness;
+    std::vector<float> position;
+    float fitness;
 };
 
 #endif //DISSERTATION_NILM_PSO_BEST_PARTICLE_H
